requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
apscheduler>=3.10.4
